requirements.txt
pyshp>=2.3.0
geopandas>=0.13.0
fiona>=1.9.0
shapely>=2.0.0
pytest>=7.0
flake8>=6.0
pandas
google-adk
matplotlib
